Apply has_attribute to the output in VTKFileReader
Fixes failing tests in simphony-mayavi due to incorrect application of assign_attribute
to a PolyData dataset.
